+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * World Map dimension where gummi ships won't use any fuel.
 * Air Soldier, Defender, Neoshadow and Novashadow.
 * Gummi shots can destroy meteors.
+* Gummi inner corner shape.
 * Creative blueprint which builds the ship rather than displaying the hologram when clicking import.
 * Option to export gummi ship blueprints into files to be able to load them as blueprints in other worlds or servers.
 * Data driven property for magic data where you can specify what the projectile does to various blocks: